requirements.txt
Django>=4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
pymongo==4.2.0
mediapipe==0.8.11
numpy==1.23.3
scikit-learn==1.1.2
psycopg2>

